fix(experiments): move eval API transform to utils and drop camel leftovers
